requirements.txt
scrapy>=2.8
beautifulsoup4>=4.11
lxml>=4.9
readability-lxml>=0.8.1
python-dateutil>=2.8
requests>=2.31
python-dotenv>=1.0
